+++ changelog.md
@@ -7,6 +7,8 @@ Here you'll find information on key changes, bug fixes and improvements made by 
 
 ### What's new in version 1906?
 * Updated to support ASDK 1.1906.0.30
+* Minor bug fixes
+* Moved backend RP resources to Tenant Space
 
 ### What's new in version 1905.1?
 * Minor bug fixes and cleanup